removed implicit Close of result
